requirements.txt
owlready
attrs
networkx
pyyaml
